[Update] IntWritable to LongWritable

+++ HW1/IMDBStudent20200994.java
@@ -12,8 +12,8 @@ import org.apache.hadoop.util.GenericOptionsParser;
 
 public class IMDBStudent20200994 {
 
-	public static class IMDBMapper extends Mapper<Object, Text, Text, IntWritable> {
-		private final static IntWritable one = new IntWritable(1);
+	public static class IMDBMapper extends Mapper<Object, Text, Text, LongWritable> {
+		private final static LongWritable one = new LongWritable(1);
 		private Text word = new Text();
 
 		public void map(Object key, Text value, Context context) throws IOException, InterruptedException {
@@ -30,12 +30,12 @@ public class IMDBStudent20200994 {
 		}
 	}
 
-	public static class IMDBReducer extends Reducer<Text,IntWritable,Text,IntWritable> {
-		private IntWritable result = new IntWritable();
+	public static class IMDBReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
+		private LongWritable result = new LongWritable();
 
-		public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-			int sum = 0;
-			for (IntWritable val : values) {
+		public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
+			long sum = 0;
+			for (LongWritable val : values) {
 				sum += val.get();
 			}
 			result.set(sum);
@@ -56,7 +56,7 @@ public class IMDBStudent20200994 {
 		job.setCombinerClass(IMDBReducer.class);
 		job.setReducerClass(IMDBReducer.class);
 		job.setOutputKeyClass(Text.class);
-		job.setOutputValueClass(IntWritable.class);
+		job.setOutputValueClass(LongWritable.class);
 		FileInputFormat.addInputPath(job, new Path(otherArgs[0]));
 		FileOutputFormat.setOutputPath(job, new Path(otherArgs[1]));
 		FileSystem.get(job.getConfiguration()).delete( new Path(otherArgs[1]), true);
